Implement schema-editing code

+++ DHS_To_Database/lib04_Update_Table_Data.py
@@ -9,10 +9,10 @@ import csv
 
 class TableDataHelper:
 
-    _MAX_COLUMN_THRESHOLD = 1000
+    _MAX_COLUMN_THRESHOLD = 500
 
     @staticmethod
-    def parse_table_name():
+    def parse_table_name(filename):
         """"For a path to a parsed data table file, return a tuple of the 
         (survey_number, country_code, file_data_type, survey_version, table_name)
         e.g. /path/to/511.CMIR71.REC43.csv -> (511, 'CM', 'IR', '71', 'REC43)"""
@@ -37,7 +37,8 @@ class TableDataHelper:
         self._populate_JSON_table_list()
         
         self._known_tables = set()
-        self._checked_column_tables = set()
+        self._modified_tables = set()
+        self._verified_tables = set()
 
     
     def _populate_JSON_table_list(self):
@@ -47,6 +48,7 @@ class TableDataHelper:
             WHERE table_schema = '{self._DATA_SCHEMA}' AND data_type = 'jsonb'""",
                 con=self._engine)
         self._json_tables = set(json_tables['table_name'])
+
 
     def _does_data_table_exist(self, table_name):
         if table_name in self._known_tables:
@@ -64,10 +66,12 @@ class TableDataHelper:
 
 
     def prepare_db_for_file(self, table_name):
-        if not self._does_data_table_exist(table_name):
-            self.create_data_table(table_name)
-        else:
-            self.check_cols_against_metadata(table_name)
+        if not table_name in self._verified_tables:
+            if not self._does_data_table_exist(table_name):
+                self.create_data_table(table_name)
+            else:
+                self.check_cols_against_metadata(table_name)
+            self._verified_tables.add(table_name)
 
 
     def create_data_table(self, table_name):
@@ -77,7 +81,7 @@ class TableDataHelper:
         maximum width currently specified for that table/column in any survey
         
         The exception if there are more than TableDataHelper._MAX_COLUMN_THRESHOLD columns 
-        specified in the metadata, in which case only columns with "id" in the name will be 
+        specified in the metadata, in which case only columns believed to be indexes will be 
         first-class columns, and a jsonb column named data will be added for storing the remainder."""
         # Note that were we just loading one single CSV we could do something like this to initialize 
         # the table.
@@ -88,9 +92,9 @@ class TableDataHelper:
             TABLESPACE pg_default;
             ALTER TABLE {self._DATA_SCHEMA}."{table_name}" OWNER to admin;"""
         if self._is_dry_run:
-            return ("Would create table with \n" + create_stmt + "\n" 
-            + " then create indices with " + self._create_or_replace_indices(table_name, is_json))
+            print (f"Would create {'JSON-type ' if is_json else ''}table {table_name} with \n" + create_stmt + "\n" )
         else:
+            print (f"Creating new {'JSON-type ' if is_json else ''}data table {table_name} with \n" + create_stmt)
             r = self._engine.execute(create_stmt)
             if is_json:
                 self._json_tables.add(table_name)
@@ -99,6 +103,29 @@ class TableDataHelper:
                 print("Executing the following to drop / recreate indices: \n" + index_sql)
                 self._engine.execute(index_sql)
         
+
+    def _col_shld_be_firstclass(self, col_name):
+        _c = col_name.lower()
+        if (("idx" in _c) or (_c.startswith("ix")) # or (_c.endswith("id"))):
+            or _c in(['surveyid','caseid','mcaseid','hhid'])):
+            # TODO check label as well, has to say "index" or "line number"
+            # e.g. 'idx94', 'ixh4', 'surveyid','caseid','mcaseid','hhid',
+            # but not e.g. 'shidioma'
+            return True
+        return False
+
+
+    def _table_should_be_json(self, table_name, n_cols):
+        if n_cols > TableDataHelper._MAX_COLUMN_THRESHOLD:
+            return True
+        is_cs =  pd.read_sql(f"""
+            SELECT EXISTS (
+                SELECT FROM {self._TABLE_SPEC_TABLE} 
+                WHERE  recordname='{table_name}'
+                AND lower(recordlabel) LIKE 'cs:%' OR lower(recordlabel) LIKE 'country specific'
+            );""", con=self._engine)['exists'][0]
+        return is_cs
+
 
     def _get_column_clauses(self, table_name):
         """Gets the columns that a new data table should have, according to the metadata tables.
@@ -109,10 +136,10 @@ class TableDataHelper:
         # list all columns that are specified for this datatable in the survey metadata (unioned 
         # set across all surveys: not all surveys will have all columns)
         whats_needed = pd.read_sql(f"""
-            SELECT lower(name) AS name, itemtype, MAX(len) AS length, MAX(start) AS start
+            SELECT lower(name) AS name, MAX(itemtype) AS itemtype, MAX(len) AS length, MAX(start) AS start
             FROM {self._TABLE_SPEC_TABLE}
             WHERE recordname='{table_name}'
-            GROUP BY name, itemtype
+            GROUP BY name
             ORDER BY start;
         """, con=self._engine)
         
@@ -120,9 +147,9 @@ class TableDataHelper:
         # every survey, the number of columns becomes very large and horribly inefficient to store 
         # (it is sparse). In these cases we store those tables with a single JSONB column for the data 
         # plus the ID/joining columns as first-class columns
-        is_json = False
-        if len(whats_needed) > TableDataHelper._MAX_COLUMN_THRESHOLD:
-            whats_needed = whats_needed[whats_needed['name'].str.contains('id')]
+        is_json = self._table_should_be_json(table_name, len(whats_needed))
+        if is_json:
+            whats_needed = whats_needed[whats_needed['name'].apply(self._col_shld_be_firstclass)]
             whats_needed.loc[len(whats_needed)] = ('data', 'JSON', '', 99999999)
             is_json=True
 
@@ -134,9 +161,11 @@ class TableDataHelper:
         def _clause_from_row(r):
             if r["itemtype"] == "JSON":
                 type_clause = "jsonb"
+                collate_clause= ""
             else:
                 type_clause = f'character varying({r["length"]})'
-            clause = f'{r["name"]} {type_clause} COLLATE pg_catalog."default"'
+                collate_clause = 'COLLATE pg_catalog."default"'
+            clause = f'{r["name"]} {type_clause} {collate_clause}'
             return clause 
 
         # convert each row in the df to a clause for use in the CREATE TABLE statement
@@ -154,7 +183,7 @@ class TableDataHelper:
         existing_indices = [i[0] for i in res.fetchall()]
 
         tbl_cols = self._engine.execute(f'SELECT * FROM dhs_data_tables."{table_name}" LIMIT 0').keys()
-        idx_fields = [c for c in tbl_cols if c.find('id') != -1]
+        idx_fields = [c for c in tbl_cols if (self._col_shld_be_firstclass(c))]
 
         drop_idx_stmts = []
         idx_stmts = []
@@ -219,17 +248,17 @@ class TableDataHelper:
         """Ensures that the varchar columns that the metadata states should be present in 
         the given table are actually present in the corresponding data table and that they 
         have the necessary width."""
-        self._ensure_column_widths(table_name)
         self._ensure_columns_presence(table_name)
+        self._ensure_column_widths(table_name)
+        
 
-
-    def _widen_column(self, table_name, column_name, req_width):
+    def _widen_column(self, table_name, column_name, req_width, cur_width=0):
         if self._is_dry_run:
             print(f"""Column {self._DATA_SCHEMA}.{table_name}.{column_name} would be widened
-                to {req_width}""")
+                to {req_width}{" (from "+str(cur_width)+")" if cur_width>0 else ""}""")
         else:
             print(f"""Widening column {self._DATA_SCHEMA}.{table_name}.{column_name} 
-                to {req_width}""")
+                to {req_width}{" (from "+str(cur_width)+")" if cur_width>0 else ""}""")
             sql = f"""ALTER TABLE {self._DATA_SCHEMA}."{table_name}" 
                     ALTER COLUMN {column_name} TYPE character varying({req_width});"""
             return self._engine.execute(sql)
@@ -244,10 +273,15 @@ class TableDataHelper:
         from a CSV of data (loaded to a dataframe), which is needed on the metadata tables 
         themselves"""
         
+        # Because this does an INNER join to information_schema.columns, columns that 
+        # are specified in the spec table but which have been stored packed in a JSON table
+        # are simply not returned by this query, which is what we need, so that we won't 
+        # attemptt to widen or check the width of a column that doesn't exist (because it's 
+        # in the JSON)
         length_specs = pd.read_sql(f"""
             SELECT 
 	            lower(s.name) AS name, MAX(s.len) as req_len, 
-	            MAX(i.character_maximum_length) as actual_len
+	            COALESCE(MAX(i.character_maximum_length),999999) as actual_len
             FROM 
                 {self._TABLE_SPEC_TABLE} s
             INNER JOIN 
@@ -258,15 +292,17 @@ class TableDataHelper:
                 'dhs_data_tables'=i.table_schema 
                 AND 
                 lower(s.name)=i.column_name
-            WHERE s.recordname={table_name}
+            WHERE s.recordname='{table_name}'
             GROUP BY s.name"""
         , con=self._engine)
         to_widen = length_specs[length_specs['req_len']>length_specs['actual_len']]
         for _, row in to_widen.iterrows():
-            self._widen_column(table_name, row['name'], row['req_len'])
+            self._widen_column(table_name, row['name'], row['req_len'], row['actual_len'])
+            self._modified_tables.add(table_name)
         
     
     def _add_varchar_column(self, table_name, column_name, req_width):
+        self._modified_tables.add(table_name)
         if self._is_dry_run:
             print(f"""Column named {column_name.lower()} would be added to 
             {self._DATA_SCHEMA}.{table_name} with width {req_width}""")
@@ -287,20 +323,22 @@ class TableDataHelper:
         is_json = table_name in self._json_tables
 
         if is_json:
+            # we only need the indexing columns to be present
             data_cols_needed = pd.read_sql(f"""
-                SELECT LOWER(name) AS name, MAX(len) AS maxlen, MAX(start) as start
+                SELECT LOWER(name) AS name, MAX(len) AS maxlen
                 FROM {self._TABLE_SPEC_TABLE}
-                WHERE recordname = '{table_name}' AND lower(name) LIKE '%id%'
-                GROUP BY name""", con=self._engine)
-            data_cols_needed.loc[len(data_cols_needed)] = ('data', 99999999, 99999999)
+                WHERE recordname = '{table_name}'
+                GROUP BY name;""", con=self._engine)
+            data_cols_needed = data_cols_needed[data_cols_needed['name'].apply(
+                self._col_shld_be_firstclass)]
+            data_cols_needed.loc[len(data_cols_needed)] = ('data',  99999999)
 
         else:
             data_cols_needed = pd.read_sql(f"""
                 SELECT LOWER(name) AS name, MAX(len) AS maxlen 
                 FROM {self._TABLE_SPEC_TABLE}
                 WHERE recordname = '{table_name}'
-                GROUP BY name
-                ORDER BY name, start""", con=self._engine)
+                GROUP BY name;""", con=self._engine)
             
         data_cols_present = pd.read_sql(f"""
             SELECT table_name, column_name, data_type 
@@ -320,31 +358,49 @@ class TableDataHelper:
     
   
     def load_table(self, table_filename, use_bulk_copy=True):
-        surveyid, _, file_type, _, table_name = parse_table_name(tbl_fn)
+        _, _, _, _, table_name = TableDataHelper.parse_table_name(table_filename)
         is_json = table_name in self._json_tables
         if is_json:
-            self._load_file_to_json_table(table_filename, surveyid, use_bulk_copy)
+            self._load_file_to_json_table(table_filename, use_bulk_copy)
         else:
             self._load_file_to_standard_table(table_filename, use_bulk_copy)
 
 
     def _load_file_to_standard_table(self, table_filename, use_bulk_copy=True):
-        surveyid, _, file_type, _, table_name = parse_table_name(tbl_fn)
+        surveyid, _, file_type, _, table_name = TableDataHelper.parse_table_name(table_filename)
         file_data = pd.read_csv(table_filename)
         file_data.columns = file_data.columns.str.lower()
         file_data['surveyid'] = surveyid
         if self._is_dry_run:
             print(f'''Would insert data from {os.path.basename(table_filename)} to 
-                {self._DATA_SCHEMA}."{table_name}"''')
+                {self._DATA_SCHEMA}."{table_name}" using {"BULK COPY" if use_bulk_copy 
+                else "standard INSERTs"}''')
         else:
-            print(f'''Inserting data from {os.path.basename(table_filename)} to 
-                {self._DATA_SCHEMA}."{table_name}"''')
-            file_data.to_sql(name=table_name, schema=self._DATA_SCHEMA,
-                con=self._engine, index=False, if_exists='append', method='multi')
+            if use_bulk_copy:
+                # use the underlying psycopg2 connection's cursor to do a streaming bulk copy insert 
+                # from an in-memory file, this is WAY faster
+                # https://stackoverflow.com/a/44181653, https://stackoverflow.com/a/44181653
+                
+                # write the DF to an in-memory buffer as TSV (tab separated)
+                print(f'''Inserting data from {os.path.basename(table_filename)} to 
+                    {self._DATA_SCHEMA}."{table_name}" using BULK COPY''')
+                buffer = io.StringIO()
+                file_data.to_csv(buffer, sep='\t', header=False, index=False)
+                qual_table = self._DATA_SCHEMA + "." + table_name
+                conn = self._engine.raw_connection()
+                cursor = conn.cursor()
+                cursor.copy_from(buffer, qual_table, sep='\t', columns=list(file_data.columns))
+                conn.commit()
+                cursor.close()
+            else:
+                print(f'''Inserting data from {os.path.basename(table_filename)} to 
+                    {self._DATA_SCHEMA}."{table_name}" using INSERTs''')
+                file_data.to_sql(name=table_name, schema=self._DATA_SCHEMA,
+                    con=self._engine, index=False, if_exists='append', method='multi')
 
 
     def _load_file_to_json_table(self, table_filename, use_bulk_copy=True):
-        surveyid, _, _, _, table_name = parse_table_name(tbl_fn)
+        surveyid, _, _, _, table_name = TableDataHelper.parse_table_name(table_filename)
         # force all columns to be read as string, otherwise pandas' type sniffer will set them 
         # to be integer where possible. For the data being stored in first class tables where the 
         # columns are all varchar, that doesn't matter as it gets cast back to string on insert. 
@@ -365,7 +421,7 @@ class TableDataHelper:
         # save it into a new column called 'data'
         file_data['data'] = file_data.to_dict('records')
         # keep only this new column (plus the "hidden" index ones)
-        file_data = file_Data[['data']]
+        file_data = file_data[['data']]
         file_data.reset_index()
         # type of data column is now a python dictionary
         # type(test['data'].iloc[0]) == dict
@@ -376,33 +432,56 @@ class TableDataHelper:
         # type of data column is now a JSON string
         # type(test['data'].iloc[0]) == str
 
-        if use_bulk_copy:
-            # use the underlying psycopg2 connection's cursor to do a streaming bulk copy insert 
-            # from an in-memory file, this is WAY faster
-            # https://stackoverflow.com/a/44181653, https://stackoverflow.com/a/44181653
+        if self._is_dry_run:
+            print(f'''Would insert data from {os.path.basename(table_filename)} to 
+                JSON table {self._DATA_SCHEMA}."{table_name}" using 
+                {"BULK COPY" if use_bulk_copy else "standard INSERTs"}''')
+        else:
+            if use_bulk_copy:
+                # use the underlying psycopg2 connection's cursor to do a streaming bulk copy insert 
+                # from an in-memory file, this is WAY faster
+                # https://stackoverflow.com/a/44181653, https://stackoverflow.com/a/44181653
+                
+                # write the DF to an in-memory buffer as TSV (tab separated)
+                print(f'''Inserting data from {os.path.basename(table_filename)} to 
+                JSON table {self._DATA_SCHEMA}."{table_name}" using BULK COPY''')
+                buffer = io.StringIO()
+                file_data.to_csv(buffer, sep='\t', header=False, index=False)
+                qual_table = self._DATA_SCHEMA + "." + table_name
+                conn = self._engine.raw_connection()
+                cursor = conn.cursor()
+                cursor.copy_from(buffer, qual_table, sep='\t', columns=list(file_data.columns))
+                conn.commit()
+                cursor.close()
             
-            # write the DF to an in-memory buffer as TSV (tab separated)
-            buffer = io.StringIO()
-            file_data.to_csv(buffer, sep='\t', header=False, index=False)
-            qual_table = self._DATA_SCHEMA + "." + table_name
-            conn = self._engine.raw_connection()
-            cursor = conn.cursor()
-            cursor.copy_from(buffer, qual_table, sep='\t', columns=list(file_data.columns))
-            conn.commit()
-            cursor.close()
+            else:
+                print(f'''Inserting data from {os.path.basename(table_filename)} to 
+                JSON table {self._DATA_SCHEMA}."{table_name}" using INSERTs''')
+                file_data.to_sql(name=table_name, schema=self._DATA_SCHEMA,
+                    con=self._engine, index=False, if_exists='append', method='multi')
             
-
-
 
     def drop_and_reload(self, tbl_fn, msg="Unknown reason"):
         """For a given data table CSV, drop the data for this surveyid from the appropriate DB table
         and then reload it."""
-        surveyid, _, file_type, _, table_name = parse_table_name(tbl_fn)
+        surveyid, _, file_type, _, table_name = TableDataHelper.parse_table_name(tbl_fn)
         self.delete_table_entries_for_survey(surveyid, table_name)
-        self.load_table(table_filename)
+        self.load_table(tbl_fn)
+
+
+    def does_survey_exist_in_table(self, surveyid, tablename):
+        return self.get_db_survey_table_rowcount(surveyid, tablename) > 0
+
 
     def get_db_survey_table_rowcount(self, surveyid, tablename):
-        pass
+        _sql = f"""SELECT count(*) nrows_db FROM {self._DATA_SCHEMA}."{table_name}"
+            WHERE surveyid={surveyid}"""
+        return self._engine.execute(sql).fetchone()[0]
+
+
+    def list_modified_tables(self):
+        return self._modified_tables
+
 
     def delete_table_entries_for_survey(self, surveyid, table_name):
         if self._is_dry_run:
@@ -414,5 +493,5 @@ class TableDataHelper:
                                     autoload=True, autoload_with=self._engine)
             cond = tablespec_db.c.surveyid == surveyid
             delete = tablespec_db.delete().where(cond)  
-            res = engine.execute(delete)
+            res = self._engine.execute(delete)
